src/backend/requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.31.1
python-multipart>=0.0.9

# AI/ML
openai>=2.14.0

# RAG (Retrieval-Augmented Generation)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
pypdf>=4.0.0

# HTTP Client
requests==2.31.0
httpx==0.27.2

# Image Processing
Pillow==10.2.0

# PaddleOCR (로컬 OCR 엔진 - PP-OCRv4)
paddlepaddle>=2.6.0
paddleocr>=2.8.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation (already included with FastAPI)
pydantic==2.12.5

# CORS
fastapi-cors==0.0.6
